requirements.txt
numpy==1.23.5
tqdm
pyyaml
librosa
soundfile
requests
compmusic==0.4
encodec
pyloudnorm
tensorboard
